# src/main/java/com/ticketing/dto/LoginRequest.java
package com.ticketing.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotNull;

public record LoginRequest(
        @NotNull @Email String email,
        @NotNull String password
) {}