# src/main/java/com/ticketing/dto/UpdateEventRequest.java
package com.ticketing.dto;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;

import java.time.Instant;

public record UpdateEventRequest(
        @NotBlank String title,
        @NotBlank String venue,
        @NotNull Instant startsAt,
        @NotNull Instant endsAt,
        @Min(1) int capacity,
        long version
) {}